fix:마이페이지 변경페이지 분리
