+++ src/pages/Main/Mypage/Mypage.tsx
@@ -1,146 +1,122 @@
-import { useState, useEffect } from 'react';
-import { useQueryClient, useMutation, useQuery } from '@tanstack/react-query';
-import Edit from './ProfileEdit/ProfileEdit';
-import View from './ProfileView/ProfileView';
-import { useAlertStore } from '../../../store/useAlertStore';
-import { updateInfo, IUserUpdateInfo, getUserInfo } from '../../../apis/user';
+import { Link } from 'react-router-dom';
+import Google from '../../../assets/images/logo/google.png';
+import Kakao from '../../../assets/images/logo/kakao.png';
+import Naver from '../../../assets/images/logo/naver.png';
+import { useMypageInfo } from '../../../hooks/useUserInfo';
 import { useUnauthorizedRedirect } from '../../../hooks/useUnauthorizedRedirect';
 import useServerErrorToast from '../../../hooks/useServerErrorToast';
 import styles from './Mypage.module.scss';
 
-interface UserInfo {
-  userName: string;
-  userEmail: string;
-  userPassword: string;
-  userSchool: string;
-  userGrade: string;
-  userMajor: string;
-  userOAuthType: string;
-  userPasswordConfirm?: string;
-}
-
 const MyPage = () => {
-  const { data, isError } = useQuery({
-    queryKey: ['user'],
-    queryFn: getUserInfo,
-  });
+  const { data, info, isError } = useMypageInfo();
 
   useUnauthorizedRedirect(data);
   useServerErrorToast(isError);
 
-  const [info, setInfo] = useState<UserInfo>({
-    userName: '',
-    userEmail: '',
-    userPassword: '',
-    userSchool: '',
-    userGrade: '',
-    userMajor: '',
-    userOAuthType: '',
-    userPasswordConfirm: '',
-  });
+  type LoginType = 'naver' | 'kakao' | 'google' | null;
 
-  const [currentMode, setCurrentMode] = useState<'view' | 'edit'>('view');
-  const queryClient = useQueryClient();
-  const addAlert = useAlertStore((state) => state.addAlert);
-
-  useEffect(() => {
-    setInfo({
-      userName: data?.object.userName || '',
-      userEmail: data?.object.userEmail || '',
-      userPassword: data?.object.userPassword || '',
-      userSchool: data?.object.userSchool || '',
-      userGrade: data?.object.userGrade || '',
-      userMajor: data?.object.userMajor || '',
-      userOAuthType: data?.object.userOAuthType || '',
-      userPasswordConfirm: '',
-    });
-  }, [data]);
-
-  const handleEditClick = () => {
-    setCurrentMode('edit');
-  };
-  const handleViewClick = () => {
-    setInfo({
-      userName: data?.object.userName || '',
-      userEmail: data?.object.userEmail || '',
-      userPassword: data?.object.userPassword || '',
-      userSchool: data?.object.userSchool || '',
-      userGrade: data?.object.userGrade || '',
-      userMajor: data?.object.userMajor || '',
-      userOAuthType: data?.object.userOAuthType || '',
-      userPasswordConfirm: '',
-    });
-    setCurrentMode('view');
+  const getLogoSrc = (type: LoginType): string => {
+    switch (type) {
+      case 'naver':
+        return Naver;
+      case 'kakao':
+        return Kakao;
+      case 'google':
+        return Google;
+      default:
+        return '';
+    }
   };
 
-  const updateMutation = useMutation({
-    mutationFn: updateInfo,
-    onSuccess: () => {
-      queryClient.invalidateQueries({ queryKey: ['user'] });
-      setCurrentMode('view');
-    },
-    onError: (error) => {
-      addAlert(
-        '정보 변경에 오류가 발생했습니다.\n 다시 시도해 주세요.',
-        'error',
-      );
-      console.error(error);
-    },
-  });
-  const handleSaveClick = () => {
-    if (info.userPassword.length < 8) {
-      addAlert('비밀번호는 최소 8자 이상이어야 합니다.', 'error');
-      return;
-    }
-    if (
-      info.userOAuthType === '' &&
-      info.userPassword.trim() !== '' &&
-      info.userPassword !== info.userPasswordConfirm
-    ) {
-      addAlert('비밀번호와 비밀번호 확인이 일치하지 않습니다.', 'error');
-      return;
-    }
+  const logoSrc = getLogoSrc(info.userOAuthType as LoginType);
+  const oauthTypeClass = info.userOAuthType && styles[info.userOAuthType];
 
-    const updateData: IUserUpdateInfo = {
-      userName: info.userName || '',
-      userPassword: info.userPassword || '',
-      userSchool: info.userSchool || '',
-      userGrade: info.userGrade || '',
-      userMajor: info.userMajor || '',
-    };
-
-    if (!info.userPassword) {
-      delete updateData.userPassword;
-    }
-    updateMutation.mutate(updateData);
-
-    setInfo({
-      userName: data?.object.userName || '',
-      userEmail: data?.object.userEmail || '',
-      userPassword: '',
-      userSchool: data?.object.userSchool || '',
-      userGrade: data?.object.userGrade || '',
-      userMajor: data?.object.userMajor || '',
-      userOAuthType: data?.object.userOAuthType || '',
-      userPasswordConfirm: '',
-    });
+  const handleLogout = () => {
+    localStorage.removeItem('token');
+    window.location.href = '/';
   };
 
   return (
     <div className={styles.wholeContainer}>
       <h1 className={styles.title}>계정 정보</h1>
       <div className={styles.profileContainer}>
-        <div className={styles.profileImgBox}></div>
-        {currentMode === 'view' ? (
-          <View info={info} onEditClick={handleEditClick} />
-        ) : (
-          <Edit
-            info={info}
-            setInfo={setInfo}
-            onSaveClick={handleSaveClick}
-            onViewClick={handleViewClick}
-          />
-        )}
+        <div className={styles.infoBox}>
+          <div className={styles.inputBox}>
+            <label className={styles.inputLabel}>
+              {logoSrc && (
+                <div className={`${styles.oauthTypeBox} ${oauthTypeClass}`}>
+                  <img
+                    className={styles.oauthTypeImg}
+                    src={logoSrc}
+                    alt={`${info.userOAuthType} Logo`}
+                  />
+                </div>
+              )}
+              계정 이메일
+              <input
+                className={`${styles.emailInput} ${!info.userOAuthType && styles.noLeftPadding}`}
+                type="text"
+                name="userEmail"
+                value={info.userEmail}
+                disabled
+              />
+            </label>
+          </div>
+          <div className={styles.inputBox}>
+            <label className={styles.inputLabel}>
+              사용자 이름
+              <input
+                className={styles.input}
+                type="text"
+                name="userName"
+                value={info.userName}
+                disabled
+              />
+            </label>
+          </div>
+          <div className={styles.schoolInputBox}>
+            <label className={styles.inputLabel}>
+              학교
+              <input
+                className={styles.schoolInput}
+                type="text"
+                name="userSchool"
+                value={info.userSchool}
+                disabled
+              />
+            </label>
+            <label className={styles.inputLabel}>
+              학년
+              <input
+                className={styles.gradeInput}
+                type="text"
+                name="userGrade"
+                value={info.userGrade}
+                disabled
+              />
+            </label>
+          </div>
+          <div className={styles.inputBox}>
+            <label className={styles.inputLabel}>
+              학과
+              <input
+                className={styles.input}
+                type="text"
+                name="userMajor"
+                value={info.userMajor}
+                disabled
+              />
+            </label>
+          </div>
+          <div className={styles.btnBox}>
+            <Link to="/home/my-page/edit" className={styles.modifyBtn}>
+              프로필 수정
+            </Link>
+            <button className={styles.logoutBtn} onClick={handleLogout}>
+              로그아웃
+            </button>
+          </div>
+        </div>
       </div>
     </div>
   );
