+++ src/pages/Main/Mypage/ProfileEdit/ProfileEdit.module.scss
@@ -1,6 +1,37 @@
 @import '../../../../styles/variables/colors.scss';
 @import '../../../../styles/variables/fonts.scss';
 
+.wholeContainer {
+  display: flex;
+  flex-direction: column;
+  width: 100%;
+  min-width: 406px;
+  gap: 20px;
+  padding: 2% 4%;
+}
+.profileContainer {
+  width: 100%;
+  min-height: 600px;
+  padding: 40px;
+  background-color: $white_FF;
+  border-radius: 8px;
+  gap: 35px;
+  display: flex;
+  justify-content: center;
+  flex-direction: column;
+  align-items: center;
+}
+.title {
+  @include SemiBold_Title_28;
+  color: $dark-font_5A;
+  padding: 7px 0;
+}
+.profileImgBox {
+  display: flex;
+  width: 100%;
+  max-width: 400px;
+}
+
 .infoBox {
   width: 100%;
   max-width: 400px;
@@ -8,6 +39,7 @@
   flex-direction: column;
   align-items: center;
   gap: 16px;
+  padding: 42px 0;
 }
 .inputBox {
   width: 100%;
@@ -69,6 +101,7 @@
   background-color: $white_FF;
   width: 24px;
   height: 24px;
+  cursor: pointer;
 }
 .gradeNum {
   color: $dark-font_33;
@@ -106,7 +139,7 @@
   width: 100%;
   display: flex;
   flex-direction: column;
-  padding: 16px 0 42px 0;
+  padding: 16px 0;
   gap: 16px;
 }
 .submitBtn {
