--- src/pages/Main/Mypage/ProfileView/ProfileView.module.scss
@@ -1,110 +1,0 @@
-@import '../../../../styles/variables/colors.scss';
-@import '../../../../styles/variables/fonts.scss';
-
-.infoBox {
-  max-width: 400px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 10px;
-}
-.inputBox {
-  width: 100%;
-  position: relative;
-}
-.inputLabel {
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  @include SemiBold_Sub_14;
-  color: $light-gray_81;
-  gap: 6px;
-}
-.input {
-  @include SemiBold_Body_16;
-  width: 100%;
-  height: 48px;
-  border: 0;
-  padding-bottom: 20px;
-  background-color: $white_FF;
-}
-.oauthTypeBox {
-  position: absolute;
-  top: 29%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  cursor: pointer;
-  border-radius: 20%;
-  width: 35px;
-  height: 35px;
-  border: none;
-}
-.google {
-  background-color: $white_FF;
-  border: 1px solid $light-bg_E5;
-}
-.kakao {
-  background-color: #f7e409;
-}
-.naver {
-  background-color: #30d341;
-}
-
-.oauthTypeImg {
-  object-fit: none;
-}
-.emailInput {
-  @include SemiBold_Body_16;
-  width: 100%;
-  height: 48px;
-  border: none;
-  padding: 5px 0px 20px 40px;
-  background-color: $white_FF;
-}
-.noLeftPadding {
-  padding-left: 0px;
-}
-.schoolInputBox {
-  width: 100%;
-  display: flex;
-  gap: 15px;
-}
-.schoolInput {
-  @include SemiBold_Body_16;
-  width: 100%;
-  height: 48px;
-  border: none;
-  padding-bottom: 20px;
-  background-color: $white_FF;
-}
-.gradeInput {
-  @include SemiBold_Body_16;
-  width: 100%;
-  height: 48px;
-  border: none;
-  padding-bottom: 20px;
-  background-color: $white_FF;
-}
-.btnBox {
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-}
-.modifyBtn {
-  height: 48px;
-  @include SemiBold_Body_16;
-  background-color: $light-bg_F2;
-  color: $dark-font_5A;
-  border: 0;
-  border-radius: 8px;
-  cursor: pointer;
-}
-.logoutBtn {
-  @include SemiBold_Body_16;
-  color: $brand-point;
-  height: 60px;
-  border: 0;
-  background-color: $white_FF;
-  cursor: pointer;
-}
